projet/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
click>=8.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
selenium>=4.10.0
webdriver-manager>=4.0.0
streamlit
plotly
psycopg2-binary>=2.9.9
